Add new IPv6 conditions
I think handling address conditions with a separate condition for IPv6 is painful, but probably the simplest approach.
It is the approach that nft takes so ultimately we have to do so on the backend.
Trying to figure out the address types automatically with just one condition and cover all the corner cases was getting too complicated.